🧪 test(代码主体无修改):
test the computation time of CPU and GPU when the ffn's weight is splited both in the DRAM and VRAM. I firstly record the computation of 2 added tensor option: MUL_MAT and AXPY, then i judge them in the add option. The  computation time is recorded in the time.txt.